Prepare intervention timing for odin

Prepare dates of intervention for use with odin. This function exists to make explicit how time changes through the model relative to the data and to odin's internal clock.

intervention_dates_for_odin(dates, change, start_date, steps_per_day)

Arguments

dates
Dates (or ISO-formatted strings for conversion with as.Date at which intervention changes.

change
Variable that is changing at each of dates.

start_date
The date to start the simulation from. Must be earlier than the first date in data.

steps_per_day
The number of steps per day

Details

If start date is after elements in dates, these will be trimmed accordinly and the final change value used as the value one day after start date.