Parmaters for explicit SEEIR model

Parmaters for explicit SEEIR model

parameters_explicit_SEEIR(
  country = NULL,
  population = NULL,
  tt_contact_matrix = 0,
  contact_matrix_set = NULL,
  R0 = 3,
  tt_R0 = 0,
  beta_set = NULL,
  time_period = 365,
  dt = 0.1,
  init = NULL,
  seeding_cases = NULL,
  prob_hosp = probs$prob_hosp,
  prob_severe = probs$prob_severe,
  prob_non_severe_death_treatment = probs$prob_non_severe_death_treatment,
  prob_non_severe_death_no_treatment = probs$prob_non_severe_death_no_treatment,
  prob_severe_death_treatment = probs$prob_severe_death_treatment,
  prob_severe_death_no_treatment = probs$prob_severe_death_no_treatment,
  p_dist = probs$p_dist,
  dur_E = 4.6,
  dur_IMild = 2.1,
  dur_ICase = 4.5,
  dur_get_ox_survive = 9.5,
  dur_get_ox_die = 7.6,
  dur_not_get_ox_survive = 9.5 * 0.5,
  dur_not_get_ox_die = 7.6 * 0.5,
  dur_get_mv_survive = 11.3,
  dur_get_mv_die = 10.1,
  dur_not_get_mv_survive = 11.3 * 0.5,
  dur_not_get_mv_die = 1,
  dur_rec = 3.4,
  hosp_bed_capacity = NULL,
  ICU_bed_capacity = NULL,
  tt_hosp_beds = 0,
  tt_ICU_beds = 0
)

Arguments

country
Character for country beign simulated. WIll be used to generate population and contact_matrix_set if unprovided. Either country or population and contact_matrix_set must be provided.

population
Population vector (for each age group). Default = NULL, which will cause population to be sourced from country

tt_contact_matrix
Time change points for matrix change. Default = 0

contact_matrix_set
Contact matrices used in simulation. Default = NULL, which will generate this based on the country.

R0
Basic Reproduction Number. Default = 3

tt_R0
Change time points for R0. Default = 0

beta_set
Alternative parameterisation via beta rather than R0. Default = NULL, which causes beta to be estimated from R0

time_period
Length of simulation. Default = 365

dt
Time Step. Default = 0.1

init
Data.frame of initial conditions. Default = NULL

seeding_cases
Initial number of cases seeding the epidemic

prob_hosp
probability of hospitalisation by age. Default = c(0.001127564, 0.000960857, 0.001774408, 0.003628171, 0.008100662, 0.015590734, 0.024597885, 0.035377529, 0.04385549, 0.058495518, 0.08747709, 0.109730508, 0.153943118, 0.177242143, 0.221362219, 0.267628264)

prob_severe
Probability of developing severe symptoms by age. Default = c(3.73755e-05, 3.18497e-05, 5.88166e-05, 0.000120264, 0.000268514, 0.000516788, 0.00081535, 0.001242525, 0.001729275, 0.002880196, 0.00598205, 0.010821894, 0.022736324, 0.035911156, 0.056362032, 0.081467057)

prob_non_severe_death_treatment
Probability of death from non severe treated infection. Default = c(0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.013361147, 0.015104687, 0.019164124, 0.027477519, 0.041762108, 0.068531658, 0.105302319, 0.149305732, 0.20349534)

prob_non_severe_death_no_treatment
Probability of death in non severe hospital inections that aren't treated

prob_severe_death_treatment
Probability of death from severe infection that is treated. Default = rep(0.5, 16)

prob_severe_death_no_treatment
Probability of death from severe infection that is not treated. Default = rep(0.95, 16)

p_dist
Preferentiality of age group receiving treatment relative to other age groups when demand exceeds healthcare capacity.

dur_E
Mean duration of incubation period (days). Default = 4.6

dur_IMild
Mean duration of mild infection (days). Default = 2.1

dur_ICase
Mean duration from symptom onset to hospitil admission (days). Default = 4.5

dur_get_ox_survive
Mean duration of oxygen given survive. Default = 5

dur_get_ox_die
Mean duration of oxygen given death. Default = 5

dur_not_get_ox_survive
Mean duration without oxygen given survive. Default = 5

dur_not_get_ox_die
Mean duration without oxygen given death. Default = 5

dur_get_mv_survive
Mean duration of ventilation given survive. Default = 7.3

dur_get_mv_die
Mean duration of ventilation given death. Default = 6

dur_not_get_mv_survive
Mean duration without ventilation given survive. Default = 7.3

dur_not_get_mv_die
Mean duration without ventilation given death. Default = 1

dur_rec
Duration of recovery after coming off ventilation. Default = 2

hosp_bed_capacity
General bed capacity. Can be single number or vector if capacity time-varies.

ICU_bed_capacity
ICU bed capacity. Can be single number or vector if capacity time-varies.

tt_hosp_beds
Times at which hospital bed capacity changes (Default = 0 = doesn't change)

tt_ICU_beds
Times at which ICU bed capacity changes (Default = 0 = doesn't change)

Value

Paramater List

Details

All durations are in days.