Plot projections against each other

Plot projections against each other

projection_plotting(
  r_list,
  scenarios,
  add_parms_to_scenarios = TRUE,
  var_select = NULL,
  replicates = FALSE,
  summarise = TRUE,
  ci = TRUE,
  q = c(0.025, 0.975),
  summary_f = mean,
  date_0 = Sys.Date(),
  x_var = "t",
  ...
)

Arguments

r_list
List of different projection runs from projections

scenarios
Character vector describing the different scenarios.

add_parms_to_scenarios
Logical. Should the parameters used for the projection runs be added to scenarios. Default = TRUE

var_select
Vector of variable names to plot (default is all)

replicates
Plot replicates

summarise
Logical, add summary line

ci
logical add confidence interval ribbon

q
Quantiles for upper and lower of interval ribbon

summary_f
Function to summarise each compartment passed to the fun argument of stat_summary

date_0
Date of time 0, if specified a date column will be added

x_var
X variable to use for plotting (default is "t", but can be set to, "date", if date_0 provided), which will cause the date to be plotted rather than time.

...
additional arguments passed to format_output