Run the deterministic explicit SEIR model

Run the deterministic explicit SEIR model

run_deterministic_SEIR_model(
  population,
  contact_matrix,
  tt_R0,
  R0_set,
  time_period,
  hosp_bed_capacity,
  ICU_bed_capacity,
  mod_gen = explicit_SEIR_deterministic,
  ...
)

Arguments

population
Population vector (for each age group). Default = NULL, which will cause population to be sourced from country

contact_matrix
Contact matrix to use in the simulation.

tt_R0
Time change points for R0

R0_set
The values of R0 to use for the simulation

time_period
Length of simulation. Default = 365

hosp_bed_capacity
General bed capacity. Can be single number of vector if capacity time-varies.

ICU_bed_capacity
ICU bed capacity. Can be single number of vector if capacity time-varies.

mod_gen
An odin model generation function. Default: `explicit_SEIR_deterministic`

...
Other arguments to pass to parameters_explicit_SEEIR

Value

Simulation output

Examples

if (FALSE) {
pop <- get_population("Afghanistan")
m <- get_mixing_matrix("Afghanistan")
run_deterministic_SEIR_model(pop$n, m, c(0, 50), c(3, 3/2), 365, 100000,
1000000)
}