Run the explicit SEEIR model

Run the explicit SEEIR model

run_explicit_SEEIR_model(
  country = NULL,
  population = NULL,
  tt_contact_matrix = 0,
  contact_matrix_set = NULL,
  R0 = 3,
  tt_R0 = 0,
  beta_set = NULL,
  time_period = 365,
  dt = 0.1,
  replicates = 10,
  init = NULL,
  seed = stats::runif(1, 0, 1e+08),
  prob_hosp = c(0.000744192, 0.000634166, 0.001171109, 0.002394593, 0.005346437,
    0.010289885, 0.016234604, 0.023349169, 0.028944623, 0.038607042, 0.057734879,
    0.072422135, 0.101602458, 0.116979814, 0.146099064, 0.176634654, 0.18),
  prob_severe = c(0.05022296, 0.05022296, 0.05022296, 0.05022296, 0.05022296,
    0.05022296, 0.05022296, 0.053214942, 0.05974426, 0.074602879, 0.103612417,
    0.149427991, 0.223777304, 0.306985918, 0.385779555, 0.461217861, 0.709444444),
  prob_non_severe_death_treatment = c(0.0125702, 0.0125702, 0.0125702, 0.0125702,
    0.0125702, 0.0125702, 0.0125702, 0.013361147, 0.015104687, 0.019164124, 0.027477519,
    0.041762108, 0.068531658, 0.105302319, 0.149305732, 0.20349534, 0.5804312),
  prob_non_severe_death_no_treatment = vapply(prob_non_severe_death_treatment * 2, min,
    numeric(1), 1),
  prob_severe_death_treatment = rep(0.5, length(prob_hosp)),
  prob_severe_death_no_treatment = rep(0.95, length(prob_hosp)),
  p_dist = rep(1, length(prob_hosp)),
  dur_E = 4.6,
  dur_IMild = 2.1,
  dur_ICase = 4.5,
  dur_get_ox_survive = 5,
  dur_get_ox_die = 5,
  dur_not_get_ox_survive = 5,
  dur_not_get_ox_die = 5,
  dur_get_mv_survive = 7.3,
  dur_get_mv_die = 6,
  dur_not_get_mv_survive = 7.3,
  dur_not_get_mv_die = 1,
  dur_rec = 2,
  hosp_bed_capacity = NULL,
  ICU_bed_capacity = NULL,
  seeding_cases = NULL
)

Arguments

country
Character for country beign simulated. WIll be used to generate population and contact_matrix_set if unprovided. Either country or population and contact_matrix_set must be provided.

population
Population vector (for each age group). Default = NULL, which will cause population to be sourced from country

tt_contact_matrix
Time change points for matrix change. Default = 0

contact_matrix_set
Contact matrices used in simulation. Default = NULL, which will generate this based on the country.

R0
Basic Reproduction Number. Default = 3

tt_R0
Change time points for R0. Default = 0

beta_set
Alternative parameterisation via beta rather than R0. Default = NULL, which causes beta to be estimated from R0

time_period
Length of simulation. Default = 365

dt
Time Step. Default = 0.5

replicates
Number of replicates. Default = 10

init
Data.frame of initial conditions. Default = NULL

seed
Random seed used for simulations. Deafult = runif(1, 0, 10000)

prob_hosp
probability of hospitalisation by age. Default = c(0.001127564, 0.000960857, 0.001774408, 0.003628171, 0.008100662, 0.015590734, 0.024597885, 0.035377529, 0.04385549, 0.058495518, 0.08747709, 0.109730508, 0.153943118, 0.177242143, 0.221362219, 0.267628264)

prob_severe
Probability of developing severe symptoms by age. Default = c(3.73755e-05, 3.18497e-05, 5.88166e-05, 0.000120264, 0.000268514, 0.000516788, 0.00081535, 0.001242525, 0.001729275, 0.002880196, 0.00598205, 0.010821894, 0.022736324, 0.035911156, 0.056362032, 0.081467057)

prob_non_severe_death_treatment
Probability of death from non severe treated infection. Default = c(0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.0125702, 0.013361147, 0.015104687, 0.019164124, 0.027477519, 0.041762108, 0.068531658, 0.105302319, 0.149305732, 0.20349534)

prob_non_severe_death_no_treatment
Probability of death in non severe hospital inections that aren't treated

prob_severe_death_treatment
Probability of death from severe infection that is treated. Default = rep(0.5, 16)

prob_severe_death_no_treatment
Probability of death from severe infection that is not treated. Default = rep(0.95, 16)

p_dist
Preferentiality of age group receiving treatment relative to other age groups when demand exceeds healthcare capacity.

dur_E
Mean duration of incubation period (days). Default = 4.6

dur_IMild
Mean duration of mild infection (days). Default = 2.1

dur_ICase
Mean duration from symptom onset to hospitil admission (days). Default = 4.5

dur_get_ox_survive
Mean duration of oxygen given survive. Default = 5

dur_get_ox_die
Mean duration of oxygen given death. Default = 5

dur_not_get_ox_survive
Mean duration without oxygen given survive. Default = 5

dur_not_get_ox_die
Mean duration without oxygen given death. Default = 5

dur_get_mv_survive
Mean duration of ventilation given survive. Default = 7.3

dur_get_mv_die
Mean duration of ventilation given death. Default = 6

dur_not_get_mv_survive
Mean duration without ventilation given survive. Default = 7.3

dur_not_get_mv_die
Mean duration without ventilation given death. Default = 1

dur_rec
Duration of recovery after coming off ventilation. Default = 2

hosp_bed_capacity
General bed capacity.

ICU_bed_capacity
ICU bed capacity.

seeding_cases
Initial number of cases seeding the epidemic

Value

Simulation output

Details

All durations are in days.

Examples

if (FALSE) {
pop <- get_population("Afghanistan")
m1 <- run_explicit_SEEIR_model(R0 = 2,
population = pop$n, dt = 1,
contact_matrix_set=contact_matrices[[1]])
}